## Linked Issues/PRs
- closes #2422 

## Description
This PR adds a benchmark target for measuring end to end query times for transaction queries. The benchmark is a simple binary that submits a bunch of transactions, and later queries these transactions by owner several times. It is intended to be used together with benchmarking tools such as [hyperfine](https://github.com/sharkdp/hyperfine).