@netrome 2024-11-15
I've rewritten the workload right now to only measure the database interactions through the `ReadDatabase` type, which isolates the effect of the multi-get implementation more than the full GraphQL interaction. This just confirms what I've already observed in the end-to-end benchmarks; the multi-get implementation is slower (for this workload).

(With the implementation from `4e7d3b357cf3b7edec95ea2859df63cf14459a35`)
```
hyperfine /tmp/batch_lookup_times_coins_reference /tmp/batch_lookup_times_coins_multi_get
Benchmark 1: /tmp/batch_lookup_times_coins_reference
  Time (mean ± σ):      6.418 s ±  0.131 s    [User: 6.258 s, System: 0.059 s]
  Range (min … max):    6.281 s …  6.647 s    10 runs

Benchmark 2: /tmp/batch_lookup_times_coins_multi_get
  Time (mean ± σ):      7.365 s ±  0.082 s    [User: 6.163 s, System: 1.105 s]
  Range (min … max):    7.239 s …  7.539 s    10 runs
```