@netrome 2024-11-15
Running the latest benches on my laptop only amplifies the difference:

Baseline:
```
hyperfine target/release/deps/end_to_end_query_times-731a676331ca9c3d
Benchmark 1: target/release/deps/end_to_end_query_times-731a676331ca9c3d
  Time (mean ± σ):     13.720 s ±  0.228 s    [User: 13.642 s, System: 0.059 s]
  Range (min … max):   13.427 s … 14.121 s    10 runs
```

With multi-get
```
hyperfine target/release/deps/end_to_end_query_times-731a676331ca9c3d
Benchmark 1: target/release/deps/end_to_end_query_times-731a676331ca9c3d
  Time (mean ± σ):     13.720 s ±  0.228 s    [User: 13.642 s, System: 0.059 s]
  Range (min … max):   13.427 s … 14.121 s    10 runs
```

I've also experimented with disabling the rocksdb block cache, which didn't make any big difference for this workload.

For this workload, I can now safely say that the multi-get implementation is strictly worse. I suspect that this is due to the overhead caused by using dynamic dispatch for the iterators, but have yet to confirm this theory.