@netrome 2024-11-15
> Reading up on multi-get, one performance benefit is reduced lock contention for concurrent reads - so I'll adjust the workload to see if I can measure anything interesting with many concurrent reads. https://github.com/facebook/rocksdb/wiki/MultiGet-Performance

Multi-get is still slower. Measured at commit `5c261ee9897282ab4e5bbee6f79a030bac5d217d`.

Reference:
```
hyperfine target/release/deps/read_view_get_coins-ff72e5522e7d0f70
Benchmark 1: target/release/deps/read_view_get_coins-ff72e5522e7d0f70
  Time (mean ± σ):      1.638 s ±  0.120 s    [User: 29.559 s, System: 0.233 s]
  Range (min … max):    1.389 s …  1.743 s    10 runs
```

Multi-get:
```
hyperfine target/release/deps/read_view_get_coins-ff72e5522e7d0f70
Benchmark 1: target/release/deps/read_view_get_coins-ff72e5522e7d0f70
  Time (mean ± σ):      1.962 s ±  0.039 s    [User: 20.858 s, System: 3.210 s]
  Range (min … max):    1.887 s …  2.019 s    10 runs
```