@netrome 2024-11-18
> What batch size are you using by default? Based on the benchmarks from the @rymnc in #2142 , the multi-get is slower for small batch size. Have you tried batch size 250/500/1000?

I've typically requested batches of 100k coins and transactions for most runs, but I have varied this parameter (`num_queries` * `utxo_count_per_block`) and haven't spotted any difference so far.

> Also, maybe you want to try to select random coins, not sequential.

Good idea, I'll try that :+1: