@netrome 2024-11-18
Tried selecting random coins (by shuffling the ID vec before each query), didn't notice any significant speed degradation in the base case.

```
hyperfine target/release/deps/read_view_get_coins-ff72e5522e7d0f70
Benchmark 1: target/release/deps/read_view_get_coins-ff72e5522e7d0f70
  Time (mean ± σ):      1.619 s ±  0.119 s    [User: 29.584 s, System: 0.247 s]
  Range (min … max):    1.415 s …  1.749 s    10 runs
```