@netrome 2024-11-18
Tried a new variant proposed by Green in a status meeting today. Only query a subset of all coins each time. With a database of 1M coins, querying 10k coins 1k times I get the following:

reference
```
hyperfine target/release/deps/read_view_get_coins-ff72e5522e7d0f70
Benchmark 1: target/release/deps/read_view_get_coins-ff72e5522e7d0f70
  Time (mean ± σ):     11.798 s ±  3.070 s    [User: 39.691 s, System: 8.286 s]
  Range (min … max):   10.156 s … 20.117 s    10 runs
```

multi-get
```
hyperfine target/release/deps/read_view_get_coins-ff72e5522e7d0f70
Benchmark 1: target/release/deps/read_view_get_coins-ff72e5522e7d0f70
  Time (mean ± σ):     12.922 s ±  3.347 s    [User: 44.348 s, System: 9.119 s]
  Range (min … max):   10.869 s … 20.437 s    10 runs
```

So multi-get still seems to be slower, but the difference seems less tangible than most recent runs. I'll see if I can find a parameter combination that changes this outcome.